refactor: meeting 패키지 만들고 meeting과 관련된 파일들 이동
