+++ android/app/src/main/java/com/woowacourse/ody/presentation/intro/IntroActivity.kt
@@ -15,7 +15,7 @@ import com.google.android.material.snackbar.Snackbar
 import com.woowacourse.ody.R
 import com.woowacourse.ody.databinding.ActivityIntroBinding
 import com.woowacourse.ody.presentation.invitecode.InviteCodeActivity
-import com.woowacourse.ody.presentation.meetinginfo.MeetingInfoActivity
+import com.woowacourse.ody.presentation.meeting.MeetingInfoActivity
 import com.woowacourse.ody.presentation.common.observeEvent
 
 class IntroActivity : AppCompatActivity() {
