+++ android/app/src/main/java/com/woowacourse/ody/presentation/meeting/MeetingInfoType.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.ody.presentation.meetinginfo
+package com.woowacourse.ody.presentation.meeting
 
 enum class MeetingInfoType {
     NAME,
