+++ android/app/src/main/java/com/woowacourse/ody/presentation/meeting/destination/MeetingDestinationFragment.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.ody.presentation.destination
+package com.woowacourse.ody.presentation.meeting.destination
 
 import android.os.Build
 import android.os.Bundle
@@ -17,8 +17,8 @@ import com.woowacourse.ody.presentation.address.AddressSearchDialog
 import com.woowacourse.ody.presentation.address.listener.AddressSearchListener
 import com.woowacourse.ody.presentation.address.model.GeoLocationUiModel
 import com.woowacourse.ody.presentation.address.model.toGeoLocation
-import com.woowacourse.ody.presentation.meetinginfo.MeetingInfoType
-import com.woowacourse.ody.presentation.meetinginfo.MeetingInfoViewModel
+import com.woowacourse.ody.presentation.meeting.MeetingInfoType
+import com.woowacourse.ody.presentation.meeting.MeetingInfoViewModel
 
 class MeetingDestinationFragment : Fragment(), AddressSearchListener {
     private var _binding: FragmentMeetingDestinationBinding? = null
