+++ android/app/src/main/java/com/woowacourse/ody/presentation/meeting/time/MeetingTimeFragment.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.ody.presentation.time
+package com.woowacourse.ody.presentation.meeting.time
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -10,8 +10,8 @@ import androidx.fragment.app.activityViewModels
 import com.google.android.material.snackbar.Snackbar
 import com.woowacourse.ody.R
 import com.woowacourse.ody.databinding.FragmentMeetingTimeBinding
-import com.woowacourse.ody.presentation.meetinginfo.MeetingInfoType
-import com.woowacourse.ody.presentation.meetinginfo.MeetingInfoViewModel
+import com.woowacourse.ody.presentation.meeting.MeetingInfoType
+import com.woowacourse.ody.presentation.meeting.MeetingInfoViewModel
 import com.woowacourse.ody.presentation.common.observeEvent
 
 class MeetingTimeFragment : Fragment() {
