+++ android/app/src/main/res/layout/activity_meeting_completion.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".presentation.completion.MeetingCompletionActivity">
+    tools:context=".presentation.meeting.complete.MeetingCompletionActivity">
 
     <ImageView
         android:layout_width="wrap_content"
