+++ android/app/src/main/res/layout/fragment_departure.xml
@@ -6,7 +6,7 @@
 
         <variable
             name="vm"
-            type="com.woowacourse.ody.presentation.meetinginfo.MeetingInfoViewModel" />
+            type="com.woowacourse.ody.presentation.meeting.MeetingInfoViewModel" />
 
         <variable
             name="addressSearchListener"
