+++ android/app/src/main/res/layout/fragment_join_nickname.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="vm"
-            type="com.woowacourse.ody.presentation.meetinginfo.MeetingInfoViewModel" />
+            type="com.woowacourse.ody.presentation.meeting.MeetingInfoViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
